Update owners when overwriting trashed file
When a trashed file is overwritten, an update is received, not an ADD.

In this case we need to make sure the same folder watcher check is
performed to add those watchers as owners of the file. If there are no
watchers, no harm done, we continue the update as before.

The differernce with ADD is that if there are no folder watchers,
there's nothing else to do and no update on the metadata is performed.