Resubmit WATCH for empty folders
If a folder watch is submitted against a folder that doesn't exist on
GCS, MS doesn't record it in the watchlist as the error is returned early.
But local-storage will record the folder in the file metadata.

This ensures that any missing folders are resubmitted by local-storage
so that a new GCS check is triggered.