Set stale when MS provides token
The second check for current version mismatch doesn't work in cases
where a watch is submitted for a file and then another watch is
submitted before the first watch has resulted in a completed download.

The first will update the version, and then the second will set
CURRENT even though the download isn't yet complete.